dynamic limiting of thread requests.
sometimes we want more requests, up to #proc (low load, anticipating a burst of work)
sometimes 1 request works the best (sustained high load, especially on a machine with a lot of cores)
So, here we try to bias the request limit once we have reasons to beleive we might be in one case or another.
It is not a precise heuristic, but appears to be statistically helpful on a range of benchmarks/machines.